// web/src/api/generated/model/keySummaryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * matthew-api
 * OpenAPI spec version: 0.1.0
 */
import type { KeySummaryResponseStatus } from './keySummaryResponseStatus';

export interface KeySummaryResponse {
  id: string;
  name: string;
  key_prefix: string;
  status: KeySummaryResponseStatus;
  created_at: string;
  last_used_at: string | null;
  never_used: boolean;
}